refactor(analysis): extract SemanticRendering from Analyzer
Move the symbol→text printers and pure SemanticDB accessors (renderType,
renderConstant, renderMethod, symbolRef, kindName, location, typeString,
parentSymbol, scopeInfos, valueType, isImplicit, rangeContains, rangeSpan)
into a stateless SemanticRendering(index) collaborator, imported back into
Analyzer so call sites stay unqualified. Shrinks Analyzer (933→840) to pure
query logic without changing behaviour.
